+++ src/lib.rs
@@ -23,6 +23,7 @@ pub mod manifest;
 pub mod mem;
 pub mod native;
 pub mod parser;
+pub mod profile;
 pub mod serve;
 pub mod token;
 pub mod transpile;